Make the label smaller to fit on screen.
As we are on the quicktag module it seems quite obvious that we are
handling quicktag. So let's just use "tag" instead of "quicktag", this
is a bit smaller and it saves spaces on the widget. This ensure that
the whole module is visible in most cases (panel size, font size...).