THIS IS A WORK IN PROGRESS!!
Took the LINGUAS file from darktable so that the lua-scripts
translations are for the same languages as darktable.

Created a data/po directory to contain the translation files.
Decided to keep the one file per script instead of one huge file.
That way the scripts can be translated one at a time if necessary
keeping the work small and manageable.

The locale directory now contains a directory for each language
contained in the LINGUAS file.  Underneath that is an LC_MESSAGES
directory and under that a scripts.mo file.

To automate maintenance of the translations I created some tools

  tools/create_potfiles.sh is a bash script run from the data/po
  directory that scans all the scripts looking for the bindtextdomain
  keyword and adding the script filename to the POTFILES.in file.

  tools/update_po_files.sh creates the translation po file for a
  script, in all the language directories, for a script if it
  doesn't have a po file.  If the po file is already there, then
  any new changes are merged with the existing po file.

  tools/fix_header.sh changes the PACKAGE, CHARSET, and Language
  fields in the po file to darktable lua-scripts, UTF-8, and the
  language code.

  tools/create_mo.sh merges the translation po files into one
  file and then creates the binary mo file from that.