Added translation to tools/script_manager.lua.
When I updated the translation files, all of them updated which
wasn't what I wanted to happen.  Added a check to only update
files when the script file is newer than the translation file or
create the translation file if it doesn't exist.

Corrected a bug where the translation file header fix ran before
the update.